--- src/main/java/com/artlable/backend/noble/command/domain/repository/NovelCharacterRepository.java
@@ -1,13 +1,0 @@
-package com.artlable.backend.noble.command.domain.repository;
-
-import com.artlable.backend.noble.command.domain.aggregate.entity.NovelCharacter;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.List;
-
-public interface NovelCharacterRepository extends JpaRepository<NovelCharacter, Long> {
-
-    List<NovelCharacter> findByCharacterOrderByCharacterNoDesc();
-
-    NovelCharacter findByCharacterNo(Long characterNo);
-}
